+++ tests/test_cli.py
@@ -188,6 +188,24 @@ def test_login_key_arg(runner, empty_netrc, local_netrc):
             generatedNetrc = f.read()
         assert DUMMY_API_KEY in generatedNetrc
 
+def test_login_onprem_key_arg(runner, empty_netrc, local_netrc):
+    onprem_key = "test-" + DUMMY_API_KEY
+    with runner.isolated_filesystem():
+        result = runner.invoke(cli.login, [onprem_key])
+        print("Output: ", result.output)
+        print("Exception: ", result.exception)
+        print("Traceback: ", traceback.print_tb(result.exc_info[2]))
+        assert result.exit_code == 0
+        with open("netrc", "r") as f:
+            generatedNetrc = f.read()
+        assert onprem_key in generatedNetrc
+
+def test_login_invalid_key_arg(runner, empty_netrc, local_netrc):
+    invalid_key = "test--" + DUMMY_API_KEY
+    with runner.isolated_filesystem():
+        result = runner.invoke(cli.login, [invalid_key])
+        assert "API key must be 40 characters long, yours was" in str(result)
+        assert result.exit_code == 1
 
 @pytest.mark.skip(reason="Just need to make the mocking work correctly")
 def test_login_anonymously(runner, monkeypatch, empty_netrc, local_netrc):
@@ -649,6 +667,7 @@ def test_local_already_running(runner, docker, local_settings):
     assert "A container named wandb-local is already running" in result.output
 
 
+@pytest.mark.skipif(platform.system() == "Windows", reason="The patch in mock_server.py doesn't work in windows")
 def test_restore_no_remote(runner, mock_server, git_repo, docker, monkeypatch):
     with open("patch.txt", "w") as f:
         f.write("test")
@@ -662,10 +681,7 @@ def test_restore_no_remote(runner, mock_server, git_repo, docker, monkeypatch):
     assert "Applied patch" in result.output
     assert "Restored config variables to " in result.output
     assert "Launching docker container" in result.output
-    docker.assert_called_with(['docker', 'run', '-e', 'LANG=C.UTF-8', '-e', 'WANDB_DOCKER=wandb/deepo@sha256:abc123', '--ipc=host', '-v',
-                            wandb.docker.entrypoint+':/wandb-entrypoint.sh', '--entrypoint', '/wandb-entrypoint.sh', '-v', os.getcwd()+
-                            ':/app', '-w', '/app', '-e',
-                            'WANDB_API_KEY=test', '-e', 'WANDB_COMMAND=python train.py --test foo', '-it', 'test/docker', '/bin/bash'])
+    docker.assert_called_with(['docker', 'run', '-e', 'LANG=C.UTF-8', '-e', 'WANDB_DOCKER=wandb/deepo@sha256:abc123', '--ipc=host', '-v', wandb.docker.entrypoint + ':/wandb-entrypoint.sh', '--entrypoint', '/wandb-entrypoint.sh', '-v', os.getcwd() + ':/app', '-w', '/app', '-e', 'WANDB_API_KEY=test', '-e', 'WANDB_COMMAND=python train.py --test foo', '-it', 'test/docker', '/bin/bash'])
 
 
 def test_restore_bad_remote(runner, mock_server, git_repo, docker, monkeypatch):
@@ -673,6 +689,7 @@ def test_restore_bad_remote(runner, mock_server, git_repo, docker, monkeypatch):
     mock_server.set_context("git", {"repo": "http://fake.git/foo/bar"})
     api = InternalApi({'project': 'test'})
     monkeypatch.setattr(cli, '_api', api)
+
     def bad_commit(cmt):
         raise ValueError()
     monkeypatch.setattr(api.git.repo, 'commit', bad_commit)