+++ wandb/lib/sparkline.py
@@ -6,6 +6,11 @@ import math
 spark_chars = u"▁▂▃▄▅▆▇█"
 
 
+# math.isfinite doesn't exist in python2, so provider our own
+def isfinite(f):
+    return not (math.isinf(f) or math.isnan(f))
+
+
 def sparkify(series):
     u"""Converts <series> to a sparkline string.
 
@@ -20,18 +25,19 @@ def sparkify(series):
     Raises ValueError if input data cannot be converted to float.
     Raises TypeError if series is not an iterable.
     """
-    series = [float(i) for i in series if not math.isnan(i)]
-    if not series:
+    series = [float(i) for i in series]
+    finite_series = [x for x in series if isfinite(x)]
+    if not finite_series:
         return u''
-    minimum = min(series)
-    maximum = max(series)
+    minimum = min(finite_series)
+    maximum = max(finite_series)
     data_range = maximum - minimum
     if data_range == 0.0:
         # Graph a baseline if every input value is equal.
-        return u''.join([spark_chars[0] for i in series])
+        return u''.join([spark_chars[0] if isfinite(x) else ' ' for x in series])
     coefficient = (len(spark_chars) - 1.0) / data_range
     return u''.join([
         spark_chars[
             int(round((x - minimum) * coefficient))
-        ] for x in series
+        ] if isfinite(x) else ' ' for x in series
     ])