+++ wandb/plots/explain_text.py
@@ -1,6 +1,7 @@
 import wandb
 from wandb import util
-from wandb.plots.utils import test_missing, test_types, encode_labels
+from wandb.plots.utils import test_missing, test_types, encode_labels, deprecation_notice
+
 
 def explain_text(text, probas, target_names=None):
         """
@@ -16,6 +17,7 @@ def explain_text(text, probas, target_names=None):
         Example:
          wandb.log({'roc': wandb.plots.ExplainText(text, probas)})
         """
+        deprecation_notice()
         eli5 = util.get_module("eli5", required="explain_text requires the eli5 library, install with `pip install eli5`")
         if (test_missing(text=text, probas=probas)):
             #and test_types(proba=proba)):