+++ wandb/plots/named_entity.py
@@ -1,6 +1,6 @@
 import wandb
 from wandb import util
-from wandb.plots.utils import test_missing, test_types, encode_labels
+from wandb.plots.utils import test_missing, test_types, encode_labels, deprecation_notice
 
 def named_entity(docs):
         """
@@ -16,6 +16,8 @@ def named_entity(docs):
         Example:
          wandb.log({'NER': wandb.plots.NER(docs=doc)})
         """
+        deprecation_notice()
+
         spacy = util.get_module("spacy", required="part_of_speech requires the spacy library, install with `pip install spacy`")
         en_core_web_md = util.get_module("en_core_web_md", required="part_of_speech requires the en_core_web_md library, install with `python -m spacy download en_core_web_md`")
         nlp = en_core_web_md.load()