+++ CHANGELOG.md
@@ -1,3 +1,35 @@
+## 0.10.7 (October 15, 2020)
+
+#### :bug: Bug Fix
+
+- Fix issue when checking for updated releases on pypi
+
+## 0.10.6 (October 15, 2020)
+
+#### :bug: Bug Fix
+
+- Make sure code saving is enabled in jupyter environments after login
+- Sweep agents have extended timeout for large sweep configs
+- Support WANDB_SILENT environment variable
+- Warn about missing python package when logging images
+- Fix wandb.restore() to apply diff patch
+- Improve artifact error messages
+- Fix loading of config-defaults.yaml and specified list of yaml config files
+
+## 0.10.5 (October 7, 2020)
+
+#### :nail_care: Enhancement
+
+- Add new custom plots: `wandb.plot.*`
+- Add new python based sweep agent: `wandb.agent()`
+
+#### :bug: Bug Fix
+
+- Console log fixes (tqdm on windows, fix close exceptions)
+- Add more attributes to the Run object (group, job_type, urls)
+- Fix sagemaker login issues
+- Fix issue where plots were not uploaded until the end of run
+
 ## 0.10.4 (September 29, 2020)
 
 #### :bug: Bug Fix